refactor: move settings repositories and notifiers to a separate settings feature folder
