+++ lib/main.dart
@@ -7,8 +7,8 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter_kids_matching_game/core/theme/app_theme.dart';
 import 'package:flutter_kids_matching_game/presentation/home_screen.dart';
 import 'package:flutter_kids_matching_game/services/storage_service.dart';
-import 'package:flutter_kids_matching_game/domain/notifiers/settings_notifier.dart';
-import 'package:flutter_kids_matching_game/presentation/settings_screen/settings_screen.dart';
+import 'package:flutter_kids_matching_game/features/settings/presentation/notifiers/settings_notifier.dart';
+import 'package:flutter_kids_matching_game/features/settings/presentation/screens/settings_screen.dart';
 import 'package:flutter_kids_matching_game/presentation/color_game_screen/color_game_screen.dart';
 import 'package:flutter_kids_matching_game/presentation/fruit_game_screen/fruit_game_screen.dart';
 import 'package:flutter_kids_matching_game/presentation/animal_game_screen/animal_game_screen.dart';
